% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\docType{package}
\name{lab04}
\alias{lab04}
\title{Package with Multiple Linear Regression}
\description{
Reference Class included:
\code{linreg}
}
\details{
For further information see \code{linreg} class documentation.
}
\author{
\itemize{
\item Farid Musayev   (gitHub:"faridmusayev", liuid:farmu437)
\item Mucahit Sahin   (gitHub:"mchtsahin", liuid:mucsa806)
\item Selen Karaduman (gitHub:"selenkaraduman", liuid:selka266)
}
}
