% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Multiple Linear Regression}
\value{
Initializes object with class "linreg" that estimates regression coefficients,
standard error, t-value, p-value and residueal standard error from a given formula and data frame
}
\description{
This is a multiple linear regression model that take two inputs (formula,data) to initialize.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula with predictors and responses (compulsory input)}

\item{\code{data}}{data frame with observations (compulsory input)}

\item{\code{data_name}}{name of data frame}

\item{\code{coefficients}}{estimated regression coefficients}

\item{\code{resids}}{residuals (difference between observed and fitted y values)}

\item{\code{predictors}}{fitted y values}

\item{\code{df}}{degrees of freedom}

\item{\code{summ}}{summary of results}

\item{\code{resid_se}}{Residual Standard Error}
}}


\examples{
linreg$new(Petal.Length~Species, data=iris)
}
